[BPK-4276] Restructure remaining F pages
