+++ docs/src/pages/FlarePage/AndroidFlarePage.js
