+++ docs/src/pages/FlarePage/FlarePage.scss
