+++ docs/src/pages/FlarePage/NativeFlarePage.js
