+++ docs/src/pages/FloatingActionButtonPage/AndroidFloatingActionButtonPage.js
