+++ docs/src/pages/FormLabelPage/FormLabelPage.js
@@ -22,7 +22,8 @@ import React from 'react';
 
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
-import WebFormLabels from '../WebFormLabelPage';
+
+import WebFormLabels from './WebFormLabelPage';
 
 const blurb = [
   <IntroBlurb>
